Init revamp obj money
